switch to timezone aware timestamps
Best practice with postgres is to use timestamps with timezone handling.
This is done by using timestamptz instead of timestamp as the column
type.

With timestamp there is no understanding of timezones and so
applications need to do all the timezone handling themselves including
being consistent in what timezone they work with from release to
release.

With timestamptz postgres will convert any timestamp that comes in to be
in UTC format and then will convert that back to the session's timezone
when read back.