Respect auto_index keys in crdb (#99)
This patch reverts a behavior introduced earlier and adds a test to prevent future regressions. With this patch, it is possible to use `change_column` on columns with foreign key references but without explicitly defined indices. Without this patch, those changes fail because CRDB needs an index on the referencing column when creating foreign key checks.

Closes #98